#2 - crate daily command

+++ database/migrations/2022_08_17_103551_create_exchanges_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('currency');
             $table->string('code');
-            $table->float('mid');
+            $table->double('mid');
         });
     }
